Move bulk finalization logic to data service
